Remove support for local installers and pruning (broken anyway*)
When I created npm-accel I was under the impression that it should be
possible to start with an empty node_modules directory, install a
non-default installer in that directory and use that installer to
install the packages that you're actually interested in. In reality I've
seen both yarn and npm-fast-install choke completely when used in this
setting. After installing them globally those problems immediately
disappeared.